add header

+++ backend/src/controllers/RecipeController.js
@@ -31,7 +31,7 @@ class RecipeController {
             title,
             ingredients,
             instructions,
-            author,
+            author: req.user.userId,  // Save the logged-in user's ID as the author
             cuisineType,
             prepTime,
             coverImage,
@@ -47,17 +47,21 @@ class RecipeController {
 
     // Update an existing recipe by ID
     static async updateRecipe(req, res) {
-        const { title, ingredients, instructions, author, cuisineType, prepTime, coverImage } = req.body;
+        const { title, ingredients, instructions, cuisineType, prepTime, coverImage } = req.body;
         try {
             const recipe = await Recipe.findById(req.params.id);
             if (!recipe) {
                 return res.status(404).json({ message: 'Recipe not found' });
             }
 
+            // Check if the logged-in user is the author of the recipe
+            if (recipe.author.toString() !== req.user.userId) {
+                return res.status(403).json({ message: 'You are not authorized to update this recipe' });
+            }
+
             recipe.title = title || recipe.title;
             recipe.ingredients = ingredients || recipe.ingredients;
             recipe.instructions = instructions || recipe.instructions;
-            recipe.author = author || recipe.author;
             recipe.cuisineType = cuisineType || recipe.cuisineType;
             recipe.prepTime = prepTime || recipe.prepTime;
             recipe.coverImage = coverImage || recipe.coverImage;
@@ -72,10 +76,17 @@ class RecipeController {
     // Delete a recipe by ID
     static async deleteRecipe(req, res) {
         try {
-            const recipe = await Recipe.findByIdAndDelete(req.params.id);
+            const recipe = await Recipe.findById(req.params.id);
             if (!recipe) {
                 return res.status(404).json({ message: 'Recipe not found' });
             }
+
+            // Check if the logged-in user is the author of the recipe
+            if (recipe.author.toString() !== req.user.userId) {
+                return res.status(403).json({ message: 'You are not authorized to delete this recipe' });
+            }
+
+            await Recipe.findByIdAndDelete(req.params.id);
             res.json({ message: 'Recipe deleted successfully' });
         } catch (error) {
             res.status(500).json({ message: error.message });
